FrontInterface/requirements.txt
arrow==0.13.1
numpy==1.16.2
plotly==2.7
django==1.11.17
django-scheduler
django-background-tasks==1.2.0
django-bootstrap3==11.0.0
django-registration-redux==2.5
django-notifications-hq==1.3.0
django-annoying==0.10.4
